Database, Api, Command, Migration
Finished up building the bases for a solid database, api, and migration pattern. I have also added the command/orchestration layer items for as we continue to develop. With this initial build we should not be able to connect everything from all the way to front (i.e. discord) all the way to the backend database. Everything should be good with this now!